% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqtheme.R
\name{arsq_draw_logo}
\alias{arsq_draw_logo}
\title{Print ggplot to device with ArsQuanta logo}
\usage{
arsq_draw_logo(plot, size = unit(2, "cm"))
}
\arguments{
\item{plot}{ggplot2 plot object}

\item{size}{default viewport size in lower right}
}
\description{
Print ggplot to device with ArsQuanta logo
}
\examples{
p <- ggplot(mtcars,aes(x = mpg,y = hp)) +
      geom_point(aes(color = cyl)) +
      labs(color = "Color",x = "MPG",y = "Horsepower") +
      theme_arsquanta()
p
arsq_draw_logo(p)
}

