% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqscale.R
\name{scale_color_arsquanta}
\alias{scale_color_arsquanta}
\title{Qualitative Color Scales for ArsQuanta}
\usage{
scale_color_arsquanta(...)
}
\description{
At the moment this simply redirects to RColorBrewer's Set1,
and so is only appropriate for qualitative scales.
}

