% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqscale.R
\name{scale_fill_arsquanta}
\alias{scale_fill_arsquanta}
\title{Qualitative Fill Scales for ArsQuanta}
\usage{
scale_fill_arsquanta(...)
}
\description{
At the moment this simply redirects to RColorBrewer's Set1,
and so is only appropriate for qualitative scales.
}

