% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aqtheme.R
\name{theme_arsquanta}
\alias{theme_arsquanta}
\title{ArsQuanta ggplot2 theme}
\usage{
theme_arsquanta(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\description{
ArsQuanta ggplot2 theme
}

